requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-google-genai
python-dotenv
pypdf
fpdf2